Add support for strict_variables (default false.)
This commit adds the strict_variables option to the twig.js tempalte
parsing process, defaulted to false. When false, if a key or variable
doesn't exists, it is assumed null. If set to true, trying to access a
key on null/undefined will throw an error. This will eventually need to
be expanded to other missing variables to stay consistent with twig-php.

Resolves #30
Resolves #27