Revert output of booleans to "true" and "false"
Pushing a string literal to a stack instead of a boolean looks like
a hack. It breaks the "sameas" tests and something else.
Modifying the "sameas" tests to support boolean->string conversion
will be hacky.

I have a fix for autoescaping already. I need it as a base.
Once it's merged I'll modify the output tag to handle output of
booleans correctly.