Fix issue with whitespace control tags where tag matching was incorrect.
Tag matching for "{{ }}{{- }}" and similar was failing due to the final
closing tag being matched rather than the earlier closing tag.

This change adds more lookahead for closing tags in the
Twig.token.findStart function, and measures the specificity and position
of the closing tags it finds versus previous matches to try and ensure
the tag sequences are balanced.