Add .editorconfig file
The `.editorconfig` file will ensure that developers will adhere to the
same coding styles and formats regardless of the editor used for
development.

By properly adjusting the settings of one's editor, style and format
violations can be prevented from happening in the first place.
As a result, one can spend less time de-linting the code.