Streamline Webpack builds
Instead of creating the server bundle and the browser bundle in two
separate Webpack builds, both bundles will be created in a single
Webpack build by having `webpack.config.js` export an array of
configuration objects.

In addition, the server build will not require transpilation as the
minimum supported NodeJS version (i.e. NodeJS 8) supports the use of
ES6 language features as is.

Finally, the `cross-env` dev dependency will be removed as the build
process no longer depends on it.