Make "js" escaped strings embeddable in JSON
The "js" escape strategy of Twig PHP [0] produces a JSON compatible escaping.
This is useful as it allows embedding JS escaped strings in JSON script.

Note: This commit replaces JSON-incompatible hexadecimal escape
sequences (\x..) with JSON-compatible unicode escape sequences (\u00..).
It also prevents the escaping of some chars that are not escaped by TwigPHP.
This fixes specific use cases, but it doesn't make the escaping fully
compatible with TwigPHP's.

[0]: https://git.io/JvEGD